thylint: option for reporting on diff lines only
The `--diff-only <revision>` option will print only matches that occur
on a line mentioned as new or changed line in the diff of the working
copy to `<revision>`. For this to work, the file names provided to the
script have to match up with the file names in the diff.

Signed-off-by: Gerwin Klein <gerwin.klein@data61.csiro.au>