Make repo-util case insensitive
GitHub and Bitbucket repo URLs are case insensitive, so some of the
manifest project names don't necessarily line up with the full (case
sensitive) repo name.

This commit brings repo-util in line with that behaviour so that it
returns valid paths for PR GITHUB_REPOSITORY references that might have
different case.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>