aws-proofs: fix cache conditions
Previously the code was trying to turn on the cache by default when a
custom name was set, but got the logic wrong (if READ was false, the
cache would also never be written).

Instead of trying to mix these, this commit simplifies the logic and
makes the name independent of reading/writing.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>