aws-proofs: add `ci_branch` input
The `run` script that is used to start proofs on an AWS VM already
supported an `INPUT_CI_BRANCH` variable. This variable controls which
branch of the ci-actions repository the VM checks out to obtain the
`steps-vm.sh` script to run.

Previously, `INPUT_CI_BRANCH` was hard-coded in `steps.sh`.
Consequently, the only way to test a change to `vm-step.sh` on a branch
was to temporarily modify `steps.sh` in the same branch.

This change adds a `ci_branch` input to the action, which allows the
branch to be controlled from workflows outside this repository. This
allows more direct testing of a proposed change to the aws-proofs
action.

Signed-off-by: Matthew Brecknell <matt@kry10.com>