aws-proofs: avoid rerunning SimplExport
This partly reverts 456f78d658. The purpose of that commit was to save
(as an artifact) the C graph-lang produced by SimplExportAndRefine, so
that it could be used in binary verification. In the case that
SimplExportAndRefine was skipped because the cached image was still
valid, the previous version forced the regeneration of the C graph-lang
by rerunning SimplExport.

In retrospect, forcing generation of the C graph-lang seems unnecessary,
since we currently do not intend to run binary verification if the CSpec
has not changed. This is because our compute resources are limited, and
we would prefer to use them for cases where the CSpec has changed.

Therefore, in this commit, we still save the C graph-lang as an
artifact, but only if it was actually produced by SimplExportAndRefine.

We remove the `simpl_export` action input that was introduced in
456f78d658, since the cost of producing the C graph-lang artifact is now
negligible.

Signed-off-by: Matthew Brecknell <matt@kry10.com>