aws-proofs: run 2 sessions in parallel
It's not entirely clear if this will really speed things up. If the VM
has n nominal cores, this will include hyper-threading, so Isabelle
will pick n/2 threads to run. With `-j 2` and without further changes
to settings, we will now potentially have all threads on the CPU busy,
which would result in slower single sessions.

But: the VMs have a fairly high CPU count (16 threads), and most
sessions will not be able to use the full 8 cores all of the time, so
there should be significant savings from not being blocked on long
sessions.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>