fetch-branch: prefer GITHUB_SHA over GITHUB_REF
GITHUB_REF is the branch name that triggered the action (e.g. master).
GITHUB_SHA is the SHA of the commit that triggered the action, but is
not always available. If it exists, prefer SHA over REF, because SHA
is stable over pushes and changes that might have happened since
the action was triggered.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>