isabelle-mirror: use old hg for upstream clone
The local hg version is fixed to a fairly old v3 to preserver git
hashes produced by the hg-git version that was used when the Isabelle
git clone was first produces.

Using `system-hg` now creates a local clone with features that are too
new for that old hg to operate on (in particular feature 'safe-share').
Over the network, old hg is still guaranteed to interoperate correctly
with recent servers, so we use that for the clone operation.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>