isabelle-mirror: hg 3.x SSL cannot clone upstream
`hg clone` does not work, because the SSL version in hg 3.x is too old
for the upstream Isabelle repo.

`hg-system clone` does work, but later leads to failure because it
produces a repo format that hg 3.x can no longer interact with.

The change instructs `hg-system` to use an older repo format.

Fixes #242

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>