await-remote-artifacts: Add new action
Add a new action that waits for artifacts to become available in another
workflow, and optionally downloads them.

This is useful when one workflow uses `repository_dispatch` to initiate
a second workflow, and the first workflow creates artifacts that are
used in the second workflow. The second workflow might need to wait for
the artifacts to become visible in the GitHub API.

This is the implementation only. Dependencies will be added in a
subsequent commit.

Signed-off-by: Matthew Brecknell <matt@kry10.com>