bv-trigger: Add action to trigger BV
When run from an l4v proof workflow, checks if there is a
`kernel-builds` artifact, and if so, sends a repository dispatch event
to graph-refine to start a binary verification run.

Previously, a called workflow was used to trigger BV runs. However, a JS
action causes less clutter in the l4v Actions tab. It's also more
efficient, because it does not need to download all artifacts to test
whether the `kernel-builds` artifact is present, and does not always
require spawning a new runner.

This is the implementation only. Dependencies will be added in a
subsequent commit.

Signed-off-by: Matthew Brecknell <matt@kry10.com>