builds.py: let mq lock be reclaimed after 30min (#337)
A typical job takes about 5-10 min. If the lock has been held for 30 min
in a single job, we assume something has gone wrong and the lock should
be released so other jobs (or humans needing the machine) can proceed.

This step is only necessary when other timeouts have failed. It is to
guard against the case where the lock release after timeout has failed,
and the post-step lock release also has failed. This can happen when the
machine queue server is temporarily unreachable on the network, and then
comes back with the lock still in place. In this case, there is nothing
the scripts here can do to release that lock.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>