docker: smaller base image for cparser-builder
The cparser-builder job is failing on GitHub, because the l4v base image
has become too large to work on GitHub runners. Use the much smaller
sel4 base image instead -- this means we have to manually install the
mlton compiler.

Also fix make invocation for the standalone parser -- this was masked
before by the fact that the l4v image comes with the parser pre-built.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>