github: always trigger actions when label present
Previously we only triggered tests such as hw-test when the
corresponding label was added. Since we now cancel older runs of the
same workflow, this no longer works, because even a run with negative
trigger condition will cancel the older workflow.

Therefore we now always trigger a new run when the test is enabled via
label.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>