link-check: do not remove tmp file at the end
The `rm -f` was masking the exit code of the link check itself, which
means that a failure there would not fail the corresponding GitHub
check. Instead of capturing and passing out the exit code, remove the
`rm -f` clean-up at the end, because the container is being discarded
anyway.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>