bashisms: remove link to versioned tar file
There are too many updates to the tar file without changes to the
bashisms checker script to go and check every time `devscripts`
gets updated and the link checker raises a failure.

Point to the directory instead, so that the link checker no longer
raises an alarm for version updates.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>