builds: remove boot failure detection
The current code did not work (never got the full log to check boot
failure on), and the mq/console scripts are already retrying on boot
failure -- more retrying is unlikely to help and will just slow down
test failure.

The hifive failures that started all this have been removed in the
meantime.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>