platforms.yml: disable HiFive P550 SMP runs for now
When doing the initial port of the P550 to seL4, I saw failures
when booting SMP but they were fixed after some improvements were
made to the locking code by Indan. However, I was testing only
in GCC and it looks like Clang has failures that must still be
investigated.

See [1] for details.

[1]: https://github.com/seL4/seL4/actions/runs/16181893559/job/45764557154

Signed-off-by: Ivan-Velickovic <i.velickovic@unsw.edu.au>